refactor(api): split oversized api modules for CI limits
Extracts helper modules from the touched API files so the changed-file size gate stays under 400 lines, while preserving behavior from origin/main. Also restores the narrower run enqueue failure semantics after the Codex adversarial review caught the regression.
