+++ cloud/apps/api/src/services/export/xlsx/pivotTable.ts
@@ -9,10 +9,14 @@
 
 import AdmZip from 'adm-zip';
 import { AppError } from '@valuerank/shared';
-
-// ============================================================================
-// TYPES
-// ============================================================================
+import {
+  colToLetter,
+  type FieldInfo,
+  generatePivotCacheDefinition,
+  generatePivotCacheRecords,
+  generatePivotTableDefinition,
+  parseCellRef,
+} from './pivotTable-xml.js';
 
 export type PivotTableConfig = {
   /** Name of the PivotTable */
@@ -36,285 +40,6 @@ export type PivotTableConfig = {
   /** Optional explicit field ordering for pivot cache values */
   fieldValueOrder?: Record<string, string[]>;
 };
-
-type FieldInfo = {
-  name: string;
-  index: number;
-  uniqueValues: string[];
-};
-
-// ============================================================================
-// XML TEMPLATES
-// ============================================================================
-
-const PIVOT_CACHE_DEFINITION_NS =
-  'xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" ' +
-  'xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships"';
-
-const PIVOT_TABLE_DEFINITION_NS =
-  'xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" ' +
-  'xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships"';
-
-// ============================================================================
-// HELPER FUNCTIONS
-// ============================================================================
-
-/**
- * Escape XML special characters.
- */
-function escapeXml(str: string): string {
-  return str
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-    .replace(/'/g, '&apos;');
-}
-
-/**
- * Parse cell reference to row and column numbers.
- */
-function parseCellRef(ref: string): { col: number; row: number } {
-  const match = ref.match(/^([A-Z]+)(\d+)$/);
-  if (match === null || match[1] === undefined || match[2] === undefined) {
-    throw new AppError(`Invalid cell reference: ${ref}`, 'XLSX_PARSE_ERROR');
-  }
-
-  const colStr = match[1];
-  const rowStr = match[2];
-
-  let col = 0;
-  for (let i = 0; i < colStr.length; i++) {
-    col = col * 26 + (colStr.charCodeAt(i) - 64);
-  }
-
-  return { col: col - 1, row: parseInt(rowStr, 10) - 1 };
-}
-
-/**
- * Convert column number to Excel column letter.
- */
-function colToLetter(col: number): string {
-  let letter = '';
-  let n = col + 1;
-  while (n > 0) {
-    const rem = (n - 1) % 26;
-    letter = String.fromCharCode(65 + rem) + letter;
-    n = Math.floor((n - 1) / 26);
-  }
-  return letter;
-}
-
-/**
- * Generate pivot cache definition XML.
- */
-function generatePivotCacheDefinition(
-  config: PivotTableConfig,
-  fields: FieldInfo[],
-  recordCount: number
-): string {
-  const cacheFields = fields
-    .map((field) => {
-      const sharedItems = field.uniqueValues
-        .map((v) => `<s v="${escapeXml(v)}"/>`)
-        .join('');
-      // Excel prefers minimal sharedItems - just count attribute
-      return `<cacheField name="${escapeXml(field.name)}" numFmtId="0">
-        <sharedItems count="${field.uniqueValues.length}">${sharedItems}</sharedItems>
-      </cacheField>`;
-    })
-    .join('\n');
-
-  // Simplified attributes to match Excel's expected format
-  return `<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<pivotCacheDefinition ${PIVOT_CACHE_DEFINITION_NS}
-    r:id="rId1"
-    refreshOnLoad="1"
-    refreshedBy="ValueRank"
-    refreshedVersion="8"
-    recordCount="${recordCount}">
-  <cacheSource type="worksheet">
-    <worksheetSource ref="${config.sourceRange}" sheet="${escapeXml(config.sourceSheet)}"/>
-  </cacheSource>
-  <cacheFields count="${fields.length}">
-    ${cacheFields}
-  </cacheFields>
-</pivotCacheDefinition>`;
-}
-
-/**
- * Generate pivot cache records XML.
- */
-function generatePivotCacheRecords(
-  fields: FieldInfo[],
-  data: string[][]
-): string {
-  const records = data
-    .map((row) => {
-      const cells = fields
-        .map((field, i) => {
-          const value = row[i] ?? '';
-          const valueIndex = field.uniqueValues.indexOf(value);
-          return `<x v="${valueIndex >= 0 ? valueIndex : 0}"/>`;
-        })
-        .join('');
-      return `<r>${cells}</r>`;
-    })
-    .join('\n');
-
-  return `<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<pivotCacheRecords ${PIVOT_CACHE_DEFINITION_NS} count="${data.length}">
-  ${records}
-</pivotCacheRecords>`;
-}
-
-/**
- * Generate pivot table definition XML.
- */
-function generatePivotTableDefinition(
-  config: PivotTableConfig,
-  fields: FieldInfo[],
-  targetLocation: string
-): string {
-  // Find field indices and their field info
-  const rowFieldInfos = config.rowFields.map((name) => {
-    const field = fields.find((f) => f.name === name);
-    if (field === undefined) throw new AppError(`Row field not found: ${name}`, 'XLSX_PARSE_ERROR');
-    return field;
-  });
-  const rowFieldIndices = rowFieldInfos.map((f) => f.index);
-
-  const colFieldInfos = config.columnFields.map((name) => {
-    const field = fields.find((f) => f.name === name);
-    if (field === undefined) throw new AppError(`Column field not found: ${name}`, 'XLSX_PARSE_ERROR');
-    return field;
-  });
-  const colFieldIndices = colFieldInfos.map((f) => f.index);
-
-  const valueFieldIndex = fields.find((f) => f.name === config.valueField)?.index;
-  if (valueFieldIndex === undefined) {
-    throw new AppError(`Value field not found: ${config.valueField}`, 'XLSX_PARSE_ERROR');
-  }
-
-  // Generate pivotFields - one for each source field
-  const pivotFields = fields
-    .map((field) => {
-      const isRowField = rowFieldIndices.includes(field.index);
-      const isColField = colFieldIndices.includes(field.index);
-      const isValueField = field.index === valueFieldIndex;
-
-      // Build attributes - a field can have both axis and dataField attributes
-      const attrs: string[] = [];
-      if (isRowField) {
-        attrs.push('axis="axisRow"');
-      } else if (isColField) {
-        attrs.push('axis="axisCol"');
-      }
-      if (isValueField) {
-        attrs.push('dataField="1"');
-      }
-      attrs.push('showAll="0"');
-
-      const items = field.uniqueValues
-        .map((_, i) => `<item x="${i}"/>`)
-        .concat(['<item t="default"/>'])
-        .join('');
-
-      return `<pivotField ${attrs.join(' ')}>
-          <items count="${field.uniqueValues.length + 1}">${items}</items>
-        </pivotField>`;
-    })
-    .join('\n');
-
-  // Row fields reference
-  const rowFields = rowFieldIndices.length > 0
-    ? `<rowFields count="${rowFieldIndices.length}">
-        ${rowFieldIndices.map((i) => `<field x="${i}"/>`).join('')}
-       </rowFields>`
-    : '';
-
-  // Row items - one <i> for each unique value plus grand total
-  // Each <i> has <x v="index"/> referencing the pivotField item
-  let rowItems = '';
-  if (rowFieldInfos.length > 0) {
-    const rowField = rowFieldInfos[0];
-    if (rowField !== undefined && rowField !== null) {
-      const items = rowField.uniqueValues
-        .map((_, i) => `<i><x v="${i}"/></i>`)
-        .concat(['<i t="grand"><x/></i>'])
-        .join('\n');
-      rowItems = `<rowItems count="${rowField.uniqueValues.length + 1}">
-        ${items}
-      </rowItems>`;
-    }
-  }
-
-  // Column fields reference
-  const colFields = colFieldIndices.length > 0
-    ? `<colFields count="${colFieldIndices.length}">
-        ${colFieldIndices.map((i) => `<field x="${i}"/>`).join('')}
-       </colFields>`
-    : '';
-
-  // Column items - one <i> for each unique value plus grand total
-  // Excel requires colItems even if empty - minimum is <colItems count="1"><i/></colItems>
-  let colItems = '<colItems count="1"><i/></colItems>';
-  if (colFieldInfos.length > 0) {
-    const colField = colFieldInfos[0];
-    if (colField !== undefined && colField !== null) {
-      const items = colField.uniqueValues
-        .map((_, i) => `<i><x v="${i}"/></i>`)
-        .concat(['<i t="grand"><x/></i>'])
-        .join('\n');
-      colItems = `<colItems count="${colField.uniqueValues.length + 1}">
-        ${items}
-      </colItems>`;
-    }
-  }
-
-  // Data fields (values) - using COUNT aggregation
-  // baseField and baseItem are required for proper value display
-  const valueLabel = config.valueFieldLabel ?? `Count of ${config.valueField}`;
-  const dataFields = `<dataFields count="1">
-    <dataField name="${escapeXml(valueLabel)}" fld="${valueFieldIndex}" subtotal="count" baseField="0" baseItem="0"/>
-  </dataFields>`;
-
-  return `<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<pivotTableDefinition ${PIVOT_TABLE_DEFINITION_NS}
-    name="${escapeXml(config.name)}"
-    cacheId="1"
-    applyNumberFormats="0"
-    applyBorderFormats="0"
-    applyFontFormats="0"
-    applyPatternFormats="0"
-    applyAlignmentFormats="0"
-    applyWidthHeightFormats="1"
-    dataCaption="Values"
-    updatedVersion="6"
-    minRefreshableVersion="3"
-    useAutoFormatting="1"
-    itemPrintTitles="1"
-    createdVersion="6"
-    indent="0"
-    outline="1"
-    outlineData="1"
-    multipleFieldFilters="0">
-  <location ref="${targetLocation}" firstHeaderRow="1" firstDataRow="${rowFieldIndices.length > 0 ? 2 : 1}" firstDataCol="1"/>
-  <pivotFields count="${fields.length}">
-    ${pivotFields}
-  </pivotFields>
-  ${rowFields}
-  ${rowItems}
-  ${colFields}
-  ${colItems}
-  ${dataFields}
-  <pivotTableStyleInfo name="PivotStyleMedium9" showRowHeaders="1" showColHeaders="1" showRowStripes="0" showColStripes="0" showLastColumn="1"/>
-</pivotTableDefinition>`;
-}
-
-// ============================================================================
-// MAIN EXPORT FUNCTION
-// ============================================================================
 
 /**
  * Add a PivotTable to an existing XLSX buffer.
